server/requirements_server.txt
grpcio
grpcio-tools
openai-whisper
google-cloud-speech
librosa
numpy
transformers
torchaudio